Authorization jwt

+++ controllers/propertiesController.js
@@ -3,15 +3,21 @@ const Rental = require('../models/rentalModel');
 
 const createProperty = async (req, res) => {
   try {
-    const newProperty = new Property(req.body);
+    const newProperty = new Property({
+      ...req.body,
+      owner: req.userId
+    });
+
     await newProperty.save();
-    console.log('Rental created:', newProperty);
+    console.log('Property created:', newProperty);
     res.status(201).json(newProperty);
   } catch (error) {
     console.error(`Error saving new property: ${error.message}`);
     res.status(500).json({ error: error.message });
   }
 };
+
+module.exports = { createProperty };
 
 const readProperty = async (req, res) => {
   try {
@@ -39,6 +45,10 @@ const updateProperty = async (req, res) => {
     if (!property) {
       return res.status(404).json({ error: "Property not found" });
     }
+    // Check if the user is the owner of the property
+    if (property.owner.toString() !== req.userId) {
+      return res.status(403).json({ error: "Access denied. You are not the owner of this property." });
+    }
 
     // Update specific fields
     for (let field in updateFields) {
@@ -60,18 +70,25 @@ const updateProperty = async (req, res) => {
 const deleteProperty = async (req, res) => {
   try {
     const { id } = req.params;
-    const rentals = await Rental.find({ property: id });
+    let property = await Property.findById(id);
+    if (!property) {
+      return res.status(404).json({ error: "Property not found" });
+    }
 
-    // Check if any rental has an 'Active' status
+    // Check if the authenticated user is the owner of the property
+    if (property.owner.toString() !== req.userId) {
+      return res.status(403).json({ error: "Access denied. You are not the owner of this property." });
+    }
+
+    // Find rentals associated with the property
+    const rentals = await Rental.find({ property: id });
     const hasActiveRentals = rentals.some(rental => rental.status === 'Active');
     if (hasActiveRentals) {
       return res.status(400).json({ error: "Cannot delete property with active rentals" });
     }
 
+    // Delete the property
     const deletedProperty = await Property.findByIdAndDelete(id);
-    if (!deletedProperty) {
-      return res.status(404).json({ error: "Property not found" });
-    }
     console.log('Property deleted:', deletedProperty);
     res.status(200).json(deletedProperty);
 
@@ -79,7 +96,7 @@ const deleteProperty = async (req, res) => {
     console.error(`Error deleting property: ${error.message}`);
     res.status(500).json({ error: error.message });
   }
-}
+};
 
 
 module.exports = {
